Pre optimizacion

+++ src/main/java/net/juligame/Debug.java
@@ -5,7 +5,7 @@ import net.juligame.classes.logic.annotations.ShowVar;
 public class Debug {
 
     @ShowVar(editable = false)
-    public int TPS = 1;
+    public double TPS = 1;
     @ShowVar(editable = false)
     public int FPS;
     @ShowVar(editable = false)
